+++ src/base/asset_file_android.cc
@@ -65,7 +65,7 @@ int AssetFile::GetSize() {
   return uncompressed_size_;
 }
 
-int AssetFile::Read(char* data, int size) {
+int AssetFile::Read(char* data, size_t size) {
   // Uncompress data into the buffer.
   int result = unzReadCurrentFile(archive_, data, size);
   return result < size ? 0 : result;
